server/requirements.txt
fastapi[standard]
sqlmodel
python-dotenv
psycopg2
clerk-backend-api
svix
python-multipart